backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic-settings==2.3.1
python-dotenv==1.0.1
celery[redis]==5.4.0
pinecone==7.3.0
langchain-openai==1.1.12
langchain-pinecone==0.2.13
langchain-community==0.4.1
langchain-huggingface==1.2.1
langchain-groq==1.1.2
sentence-transformers==3.0.1
playwright==1.44.0
beautifulsoup4==4.12.3
httpx==0.27.0
langchain==1.2.14
langchain-core==1.2.23
langchain-text-splitters==1.1.1
